requirements.txt
anyio==4.6.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorlog==6.8.2
cryptography==43.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.11
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
lomond==0.3.3
lxml==5.3.0
numpy==1.26.4
packaging==24.1
pandas==2.1.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tabulate==0.9.0
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
